Move fetchMsigState
