+++ utils/fetchMsigState.test.ts
@@ -2,13 +2,13 @@ jest.mock('@glif/filecoin-rpc-client')
 import { FilecoinNumber } from '@glif/filecoin-number'
 import { convertAddrToPrefix } from '@glif/react-components'
 
-import fetchMsigState from './fetchMsigState'
+import { fetchMsigState } from './fetchMsigState'
 
 import {
   MULTISIG_ACTOR_ADDRESS,
   MULTISIG_SIGNER_ADDRESS,
   MULTISIG_SIGNER_ADDRESS_2
-} from '../../test-utils'
+} from '../test-utils'
 
 jest
   .spyOn(require('../../apolloClient'), 'createApolloClient')
