--- utils/msig/index.js
@@ -1,1 +1,0 @@
-export { default as fetchMsigState } from './fetchMsigState'
